refactor: moved game AI into separate file
