+++ src/components/games/tictactoe/TicTacToe.tsx
@@ -8,23 +8,12 @@ import {
 } from "@/components/hooks/useGameStatistics";
 import { GAMES_DAILY_COIN_LIMIT } from "@/utils/constants";
 import { GameName, GameResult } from "@/types/games";
-import { TICTACTOE_MINIMAX_PROBABILITY } from "@/constant/tictactoeConstants";
-
-type Board = (string | null)[];
-type Difficulty = "easy" | "normal" | "expert";
-
-function makeRandomMove(currentBoard: Board): Board {
-  const emptySquares = currentBoard
-    .map((val, idx) => (val === null ? idx : null))
-    .filter((val) => val !== null) as number[];
-
-  if (emptySquares.length === 0) return currentBoard;
-
-  const randomIndex = Math.floor(Math.random() * emptySquares.length);
-  const newBoard = [...currentBoard];
-  newBoard[emptySquares[randomIndex]] = "O";
-  return newBoard;
-}
+import {
+  calculateWinner,
+  makeAIMove,
+  type Board,
+  type Difficulty,
+} from "./tictactoeAI";
 
 export default function TicTacToe({
   setSpeechText,
@@ -154,94 +143,6 @@ export default function TicTacToe({
     setSpeechText(message);
   }, [difficulty, gameState, setSpeechText]);
 
-  const calculateWinner = (squares: Board): string | null => {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (
-        squares[a] &&
-        squares[a] === squares[b] &&
-        squares[a] === squares[c]
-      ) {
-        return squares[a];
-      }
-    }
-    return null;
-  };
-
-  // for the expert ai
-  const minimax = (currentBoard: Board, isMaximizing: boolean): number => {
-    const winner = calculateWinner(currentBoard);
-    if (winner === "O") return 10;
-    if (winner === "X") return -10;
-    if (currentBoard.every((square) => square !== null)) return 0;
-
-    if (isMaximizing) {
-      let bestScore = -Infinity;
-      for (let i = 0; i < currentBoard.length; i++) {
-        if (currentBoard[i] === null) {
-          const newBoard = [...currentBoard];
-          newBoard[i] = "O";
-          const score = minimax(newBoard, false);
-          bestScore = Math.max(score, bestScore);
-        }
-      }
-      return bestScore;
-    } else {
-      let bestScore = Infinity;
-      for (let i = 0; i < currentBoard.length; i++) {
-        if (currentBoard[i] === null) {
-          const newBoard = [...currentBoard];
-          newBoard[i] = "X";
-          const score = minimax(newBoard, true);
-          bestScore = Math.min(score, bestScore);
-        }
-      }
-      return bestScore;
-    }
-  };
-
-  const getBestMove = (currentBoard: Board): number => {
-    let bestScore = -Infinity;
-    let bestMove = 0;
-
-    for (let i = 0; i < currentBoard.length; i++) {
-      if (currentBoard[i] === null) {
-        const newBoard = [...currentBoard];
-        newBoard[i] = "O";
-        const score = minimax(newBoard, false);
-        if (score > bestScore) {
-          bestScore = score;
-          bestMove = i;
-        }
-      }
-    }
-    return bestMove;
-  };
-
-  const makeAIMove = (currentBoard: Board): Board => {
-    const probability = TICTACTOE_MINIMAX_PROBABILITY[difficulty];
-    const useMinimmax = Math.random() < probability;
-
-    if (useMinimmax) {
-      const bestMove = getBestMove(currentBoard);
-      const newBoard = [...currentBoard];
-      newBoard[bestMove] = "O";
-      return newBoard;
-    } else {
-      return makeRandomMove(currentBoard);
-    }
-  };
-
   const handleGameWin = async () => {
     if (!userId || isProcessingWin) return;
 
@@ -321,7 +222,7 @@ export default function TicTacToe({
     // 2 second delay
     aiMoveTimeoutRef.current = setTimeout(() => {
       aiMoveTimeoutRef.current = null;
-      const aiBoard = makeAIMove(newBoard);
+      const aiBoard = makeAIMove(newBoard, difficulty);
       const aiWinner = calculateWinner(aiBoard);
       if (aiWinner === "O") {
         if (userId) {
